Fix sizeof args bug & cleanup
Turns out uint8_t is passed by reference by default since it's just a
typedef for char. As a result the length of args cannot be computed in
runCommand. The caller is now responsible for computing the number of
args and then passing them to runCommand

openCameraFd now returns the result and writes the file descriptor in
the fd pointer.

Fix an inverted if statement in runComamnd

Fix inverted read logic in snapShotToFile function