Use system serial calls and cleanup
The issues related to reading photos are a result of an issue in the
Legato le_tty interface. Opening a serial connection using all system
calls solves this problem nicely. Thanks to wiringPi for providing some
code for this.

All the log statements are cleaned up. The stream functions could use
some feedback to let the developer know things are still alive.

There are also doc comments for each function except some of the motion
detection stuff that I have not quite figured out yet.